requirements.txt
easyocr==1.7.2
Flask==3.1.0
numpy==2.1.3
openai==1.55.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.2
google-auth==2.36.0
google-cloud-bigquery==3.27.0
google-auth-oauthlib==1.2.1
